Implement the most basic functionality
This app will identify if the youtube
comments on a video is disabled or not and only then create a new video.
To do so I used youtube API and worked till late
night to create this system.
It is nothing big, I was just not sure what-to-use & how to use it.
This took a lot of time but it was worth it